requirements.txt
numpy==1.19.5
tensorflow-cpu==2.6.0
flask==2.0.1
werkzeug==2.0.1
gunicorn==20.1.0
scikit-learn==0.22.2.post1
Pillow==8.3.1